ZMX-#66 moved custom exception in exceptions package
