+++ src/main/java/dev/exceptions/CustomException.java
@@ -1,4 +1,4 @@
-package dev.utils;
+package dev.exceptions;
 
 public class CustomException extends RuntimeException {
     public CustomException(String message, Throwable cause) {
